// frontend/src/app/features/matches/models/matchesPaginated.model.ts
import { Match } from './match.model';

export interface MatchesPaginated {
  content: Match[];
  totalElements: number;
  pageable: {
    pageNumber: number;
    pageSize: number;
    sort: {
      empty: boolean;
      sorted: boolean;
      unsorted: boolean;
    };
    offset: number;
    paged: boolean;
    unpaged: boolean;
  };
}